Add a quick & dirty dirtree resource type
This allows you to declare a resource type that simply ensures that a
directory exists on the client. You cannot manage pwermissions or
ownership or anything, that's what the file type is for. It's just for
those edge cases when you just need to make sure a directory exists and
you don't have the liberty of managing the full path.